[ce41xx] Changed the build system around some.
Now we build in /opt/xbmc-cex by default. This way we can share paths
between our build system and target system without worrying about rpath
and friends. Also, it is now very easy to mount /opt/xbmc-cex via a
separate nfs share, so that no files have to be copied in order to test
a build.