[cex] add gconv-modules too the rootfs
for some reason, intel's sdk strips gconv-modules from libc's install.
This means that anything not preloading libiconv directly (or built with
its header and linked) will fail at runtime, as libc's is used by default.

This file was taken from Intel's SDK and is the one that would be
installed were it not stripped out.

The other option would be for us to link to libiconv, making sure that
its path comes before libc's in [C|LD]flags. For the sake of not
rattling upstream with a --with-libiconv, we'll just use libc's for now.