added: XB_FMT_R8G8B8 format.
This allows us to use rgb for textures with no alpha. GL converts to rgba for
the internal texture format, so the upload should (i think?) still be
byte-aligned.

For opaque textures in GLES this also saves a swizzle in the shader.
Before, we uploaded bgra data and converted to rgba.

The PVR glsl compiler shows a reduction from 3 operations to 1 when blending,
and from 2 to 1 when not.