changed: improved gles performance by avoiding bgra->rgba convert for textures
Previously, bgra textures were uploaded as rgba then converted back in
shaders. Instead, check for the existance of BGRA extensions and convert
on upload instead. If no extension is found, convert in software before
uploading.

This allows us to remove the swizzle in our shaders, which amounts to a
significant performance gain.