d-r: add new mode and set it as default
added: DIRTYREGION_SOLVER_FILL_VIEWPORT_ON_CHANGE
This solver (mode 3) always marks the entire screen as dirty. The difference
between this and mode 0 is that it stops rendering when there are no dirty
regions left.

This should be a safe default as it does not require preserved backbuffers,
but it does require that we stop flipping. See next commits.