jpeg: be sure the current texture is big enough
When using the large texture manager, we can load a scaled version of the
original as long as it's bigger than the desired size.
This applies mainly to jpegs, which can be partially decoded (losslessly
depending on your definition) much faster than decoding the entire image and
scaling it down.