atlas: fix refcounting
The refcount was poisoned by the GetTexture() when the subtexture was added.
With that fixed, we're able to use the parent refcount reliably.

As a result, the parent texture now tears down correctly. Its unload is skipped
to avoid re-uploading it for each screen-change. This should probably go on
some type of timer instead? Either way, we'll need some way to force unload for
switching skins.