[droid] add a way to specify forbidden players based on playerfactory rules
Use the "forbiddenplayers" attribute in a rule in playercorefactory.xml to
specify that a player should NOT be used for a specific rule. Multiple comma-
separated players can be specified. Forbidden rules trump all others.

Example:
.nsv is an audio file, but paplayer can't play it. So the current rule forces
it to dvdplayer instead. This is a problem when a third (default) player is
introduced that may handle .nsv with no problems. The solution then is to
forbid paplayer instead.