depends: add a test-dependencies target for verifying our dependency tree
For each lib, delete staging, make distclean in depends, build the lib's
dependencies, then build the lib. This effectively walks down the entire tree
from each and every leaf. Libs will be rebuilt multiple times. It will take
FOREVER. But if it is successful, it's almost certain that our dependency
tree is stable.